+++ mirage/config.ts
@@ -66,13 +66,9 @@ export default function(this: Server) {
 
     // SHARE-powered search endpoints
     this.urlPrefix = shareBaseUrl;
-    this.namespace = '/api/v3/';
+    this.namespace = '/trove/'; // /api/v3/ works as well, but /trove/ is the preferred URL
     this.get('/index-card-search', cardSearch);
     this.get('/index-value-search', valueSearch);
-    // this.get('/index-card/:id', Detail);
-    this.get('/index-card-searches', cardSearch);
-    this.get('/index-value-searches', valueSearch);
-    // this.get('/index-cards/:id', Detail);
 
     this.urlPrefix = osfUrl;
     this.namespace = '/api/v1/';
@@ -354,6 +350,11 @@ export default function(this: Server) {
     osfResource(this, 'preprint');
     this.post('/preprints', createPreprint);
 
+    this.get('/preprints/:id', (schema, request) => {
+        const id = request.params.id;
+        return schema.preprints.find(id);
+    });
+
     osfNestedResource(this, 'preprint', 'contributors', {
         path: '/preprints/:parentID/contributors/',
         defaultSortKey: 'index',
@@ -371,6 +372,17 @@ export default function(this: Server) {
         path: '/preprints/:parentID/files/',
         defaultSortKey: 'index',
         relatedModelName: 'file',
+    });
+
+    osfNestedResource(this, 'preprint', 'affiliatedInstitutions', {
+        path: '/preprints/:parentID/institutions/',
+        defaultSortKey: 'index',
+        relatedModelName: 'institution',
+    });
+
+    osfToManyRelationship(this, 'preprint', 'affiliatedInstitutions', {
+        only: ['related', 'update', 'add', 'remove'],
+        path: '/preprints/:parentID/relationships/institutions',
     });
 
     this.put('/preprints/:parentID/files/:fileProviderId/upload', uploadToRoot); // Upload to file provider